tech(cleanup): refactor search test before starting to write mobile tests
