+++ test/test_helper.rb
@@ -8,6 +8,7 @@ require 'rails/test_help'
 require 'capybara-screenshot/minitest'
 require 'support/api_test_helpers'
 require 'support/third_party_test_helpers'
+require 'support/search_internship_offer_helpers'
 require 'support/email_spam_euristics_assertions'
 require 'support/organisation_form_filler'
 require 'support/internship_offer_info_form_filler'
